# python-flask-server/openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: mduby@broadinstitute.org
    name: Genetics Data Provider
    x-role: responsible organization
  description: Genetics Data Provider for NCATS Biomedical Translator Reasoners
  license:
    name: MIT License
    url: https://github.com/broadinstitute/genetics-kp-dev/blob/master/LICENSE
  termsOfService: https://github.com/broadinstitute/genetics-kp-dev/blob/master/terms_of_service.md
  title: Genetics Data Provider for NCATS Biomedical Translator Reasoners
  version: 1.1.0.0
  x-translator:
    component: KP
    team:
    - Genetics Data Provider
    biolink-version: 1.8.0
    externalDocs:
      description: The values for component and team are restricted according to this
        external JSON schema. See schema and examples at url
      url: https://github.com/NCATSTranslator/translator_extensions/blob/\ production/x-translator/
  x-trapi:
    version: 1.1.0-beta
    externalDocs:
      description: The values for version are restricted according to the regex in
        this external JSON schema. See schema and examples at url
      url: https://github.com/NCATSTranslator/translator_extensions/blob/\ production/x-trapi/
...
servers:
- description: TRAPI production service for the Translator Genetics Data Provider
  url: https://translator.broadinstitute.org/genetics_provider/trapi/v1.1
...
paths:
...
  /query:
    post:
      operationId: query_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Query'
        description: Query information to be submitted
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
          description: OK. There may or may not be results. Note that some of the
            provided identifiers may not have been recognized.
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request. The request is invalid according to this OpenAPI
            schema OR a specific identifier is believed to be invalid somehow (not
            just unrecognized).
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Internal server error.
        "501":
          content:
            application/json:
              schema:
                type: string
          description: Not implemented.
      summary: Query reasoner via one of several inputs
      tags:
      - query
      x-openapi-router-controller: openapi_server.controllers.query_controller
components:
  schemas:
    Query:
      additionalProperties: true
      description: The Query class is used to package a user request for information.
        A Query object consists of a required Message object with optional additional
        properties. Additional properties are intended to convey implementation-specific
        or query-independent parameters. For example, an additional property specifying
        a log level could allow a user to override the default log level in order
        to receive more fine-grained log information when debugging an issue.
      example:
        message:
          query_graph:
            edges:
              e00:
                subject: n00
                predicate: biolink:affects
                object: n01
            nodes:
              n00:
                id: PUBCHEM.COMPOUND:2244
                category: biolink:ChemicalSubstance
              n01:
                category: biolink:Gene
        log_level:
          description: The least critical level of logs to return
          allOf:
          - $ref: '#/components/schemas/LogLevel'
          nullable: true
      properties:
        message:
          $ref: '#/components/schemas/Message'
      required:
      - message
      type: object
      x-body-name: request_body
    Response:
      additionalProperties: true
      description: The Response object contains the main payload when a TRAPI query
        endpoint interprets and responds to the submitted query successfully (i.e.,
        HTTP Status Code 200). The message property contains the knowledge of the
        response (query graph, knowledge graph, and results). The status, description,
        and logs properties provide additional details about the response.
      example:
        description: Success. 42 results found.
        message:
          knowledge_graph: ""
          results:
          - edge_bindings:
              key:
              - id: id
              - id: id
            node_bindings:
              key:
              - id: id
              - id: id
          - edge_bindings:
              key:
              - id: id
              - id: id
            node_bindings:
              key:
              - id: id
              - id: id
          query_graph: ""
        logs:
        - code: code
          level: ""
          message: message
          timestamp: 2020-09-03T18:13:49Z
        - code: code
          level: ""
          message: message
          timestamp: 2020-09-03T18:13:49Z
        status: Success
      properties:
        message:
          $ref: '#/components/schemas/Message'
        status:
          description: One of a standardized set of short codes, e.g. Success, QueryNotTraversable,
            KPsNotAvailable
          example: Success
          nullable: true
          type: string
        description:
          description: A brief human-readable description of the outcome
          example: Success. 42 results found.
          nullable: true
          type: string
        logs:
          description: Log entries containing errors, warnings, debugging information,
            etc
          items:
            $ref: '#/components/schemas/LogEntry'
          nullable: true
          type: array
      required:
      - message
      type: object
    Message:
      additionalProperties: false
      description: 'The message object holds the main content of a Query or a Response
        in three properties: query_graph, results, and knowledge_graph. The query_graph
        property contains the query configuration, the results property contains any
        answers that are returned by the service, and knowledge_graph property contains
        lists of edges and nodes in the thought graph corresponding to this message.
        The content of these properties is context-dependent to the encompassing object
        and the TRAPI operation requested.'
      example:
        knowledge_graph: ""
        results:
        - edge_bindings:
            key:
            - id: id
            - id: id
          node_bindings:
            key:
            - id: id
            - id: id
        - edge_bindings:
            key:
            - id: id
            - id: id
          node_bindings:
            key:
            - id: id
            - id: id
        query_graph: ""
      properties:
        results:
          description: List of all returned Result objects for the query posed
          items:
            $ref: '#/components/schemas/Result'
          nullable: true
          type: array
        query_graph:
          allOf:
          - $ref: '#/components/schemas/QueryGraph'
          description: QueryGraph object that contains a serialization of a query
            in the form of a graph
          nullable: true
        knowledge_graph:
          allOf:
          - $ref: '#/components/schemas/KnowledgeGraph'
          description: KnowledgeGraph object that contains lists of nodes and edges
            in the thought graph corresponding to the message
          nullable: true
      type: object
    LogEntry:
      additionalProperties: true
      description: The LogEntry object contains information useful for tracing and
        debugging across Translator components.  Although an individual component
        (for example, an ARA or KP) may have its own logging and debugging infrastructure,
        this internal information is not, in general, available to other components.
        In addition to a timestamp and logging level, LogEntry includes a string intended
        to be read by a human, along with one of a standardized set of codes describing
        the condition of the component sending the message.
      example:
        code: code
        level: ""
        message: message
        timestamp: 2020-09-03T18:13:49Z
      properties:
        timestamp:
          description: Timestamp in ISO 8601 format
          example: 2020-09-03T18:13:49Z
          format: date-time
          nullable: true
          type: string
        level:
          allOf:
          - $ref: '#/components/schemas/LogLevel'
          nullable: true
        code:
          description: One of a standardized set of short codes e.g. QueryNotTraversable,
            KPNotAvailable, KPResponseMalformed
          nullable: true
          type: string
        message:
          description: A human-readable log message
          nullable: true
          type: string
      type: object
    LogLevel:
      description: Logging level
      enum:
      - ERROR
      - WARNING
      - INFO
      - DEBUG
      type: string
    Result:
      additionalProperties: true
      description: A Result object specifies the nodes and edges in the knowledge
        graph that satisfy the structure or conditions of a user-submitted query graph.
        It must contain a NodeBindings object (list of query graph node to knowledge
        graph node mappings) and an EdgeBindings object (list of query graph edge
        to knowledge graph edge mappings).
      example:
        edge_bindings:
          key:
          - id: id
          - id: id
        node_bindings:
          key:
          - id: id
          - id: id
      properties:
        node_bindings:
          additionalProperties:
            items:
              $ref: '#/components/schemas/NodeBinding'
            type: array
          description: The dictionary of Input Query Graph to Result Knowledge Graph
            node bindings where the dictionary keys are the key identifiers of the
            Query Graph nodes and the associated values of those keys are instances
            of NodeBinding schema type (see below). This value is an array of NodeBindings
            since a given query node may have multiple knowledge graph Node bindings
            in the result.
          type: object
        edge_bindings:
          additionalProperties:
            items:
              $ref: '#/components/schemas/EdgeBinding'
            type: array
          description: The dictionary of Input Query Graph to Result Knowledge Graph
            edge bindings where the dictionary keys are the key identifiers of the
            Query Graph edges and the associated values of those keys are instances
            of EdgeBinding schema type (see below). This value is an array of EdgeBindings
            since a given query edge may resolve to multiple knowledge graph edges
            in the result.
          type: object
      required:
      - edge_bindings
      - node_bindings
      type: object
    NodeBinding:
      additionalProperties: true
      example:
        id: id
      properties:
        id:
          description: A Compact URI, consisting of a prefix and a reference separated
            by a colon, such as UniProtKB:P00738. Via an external context definition,
            the CURIE prefix and colon may be replaced by a URI prefix, such as http://identifiers.org/uniprot/,
            to form a full URI.
          externalDocs:
            url: https://www.w3.org/TR/2010/NOTE-curie-20101216/
          type: string
      required:
      - id
      type: object
...
    QueryGraph:
      additionalProperties: true
      description: A graph representing a biomedical question. It serves as a template
        for each result (answer), where each bound knowledge graph node/edge is expected
        to obey the constraints of the associated query graph element.
      properties:
        nodes:
          additionalProperties:
            $ref: '#/components/schemas/QNode'
          description: The node specifications. The keys of this map are unique node
            identifiers and the corresponding values include the constraints on bound
            nodes.
          type: object
        edges:
          additionalProperties:
            $ref: '#/components/schemas/QEdge'
          description: The edge specifications. The keys of this map are unique edge
            identifiers and the corresponding values include the constraints on bound
            edges, in addition to specifying the subject and object QNodes.
          type: object
      required:
      - edges
      - nodes
      type: object
...